move path json reader
