+++ src/test/java/me/saro/commons/JsonReaderTest.java
@@ -5,6 +5,8 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
+import me.saro.commons.json.JsonReader;
+
 public class JsonReaderTest {
 
     @Test
